more warnings fixes
git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code/HxCFloppyEmulator/libhxcfe/trunk@1834 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69

+++ sources/xml_disk/xml_disk.c
@@ -1233,7 +1233,7 @@ int32_t hxcfe_isMatchingXmlFloppy ( HXCFE_XMLLDR* xmlfb_ctx, char * filename, ui
 	// Check file size
 	if( ad->file_size != -1 )
 	{
-		if( buffersize != ad->file_size )
+		if( buffersize != (uint32_t)ad->file_size )
 		{
 			return 0;
 		}
